Activity-page
Users can set their activities on the activity page, which also gives them the option to select from a list of suggested activities.This page will send feedback based on the activities you are expected to complete. If the user completes the activities, good feedback will be sent; if not, negative feedback will be sent, urging the user to complete enough activities.